+++ core/src/main/java/com/froso/ufp/modules/optional/sms/service/LowLevelSMSService.java
@@ -1,15 +1,5 @@
 package com.froso.ufp.modules.optional.sms.service;
 
-import com.froso.ufp.core.domain.documents.*;
-import com.froso.ufp.modules.core.client.service.*;
-import com.froso.ufp.modules.optional.messaging.model.messaging.*;
-import com.froso.ufp.modules.optional.sms.model.*;
-import com.froso.ufp.modules.optional.sms.model.sms.*;
-import org.springframework.data.domain.*;
-import org.springframework.stereotype.*;
-
-import java.util.*;
-
 @Service
 public class LowLevelSMSService extends AbstractClientRefService<QueueSms> {
 
@@ -44,11 +34,11 @@ public class LowLevelSMSService extends AbstractClientRefService<QueueSms> {
     }
 
     public List<QueueSms> findAllNew(Pageable p) {
-        return findByKeyValue("info.status", "WAITING_TO_SEND");
+        return findByKeyValue("info.status", "=" + MessageStatusEnum.WAITING_TO_SEND);
     }
 
     public List<QueueSms> findAllErrornous(Pageable p) {
-        return findByKeyValue("info.status", "ERROR");
+        return findByKeyValue("info.status", "=" + "ERROR");
 
     }
 }